Ensure flat namespace applies
The plugin now successfully generates a site with a flat namespace as
expected.

Two things: When Document.url or Page.url is called, that freezes the
value. So we use WorkingUrl to create a URL value that we can work with
long before one of those other methods gets called, so we can set the
permalink before then.

Second: The sidebar was always getting generated as if the namespace was
always hierarchical. Now we leave out the parent link prefix while
generating `_includes/sidebar-children.html` when `flat_namespace:` is
set to true in the config.

`test/tags_test.rb` now requires the top-level `jekyll` module to fix an
error whereby running it standalone via `bundle exec rake test
TEST=test/tags_test.rb` would result in `jekyll/tags/include` raising an
uninitialized constant error.